requirements.txt
Django==4.0
psycopg2-binary==2.9.2
Pillow==8.4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
markdown==3.3.6
django-filter==21.1
djangorestframework-simplejwt==5.0.0